+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Fix bug where EVENTARC_CLOUD_EVENT_SOURCE environment variable was correctly set for some functions. (#5597)
 - Adds 2nd gen firestore triggers to firebase deploy (#5592).
 - Adds Extension emulator support for system params.
+- Adds `database:import` command for non-atomic imports (#5396).